requirements.txt
PyQt6>=6.5.0
mistune>=3.0.0
WeasyPrint>=60.0
Jinja2>=3.1.0
PyYAML>=6.0
py2app>=0.28
watchdog>=3.0.0  # For auto-reloading preview

